+++ src/Particle.h
@@ -10,12 +10,12 @@
 #include <utility>
 
 struct Particle {
-    Eigen::Vector3d x; // position
-    Eigen::Vector3d p; // estimate for new position
-    Eigen::Vector3d v; // velocity
-    Eigen::Vector3d F; // external forces
-    double m; // mass
-    double w; // weight w = 1/m
+    Eigen::Vector3f x; // position
+    Eigen::Vector3f p; // estimate for new position
+    Eigen::Vector3f v; // velocity
+    Eigen::Vector3f F; // external forces
+    float m; // mass
+    float w; // weight w = 1/m
 };
 
 #endif //PARTICLE_H
